refactor socket module
