+++ backend/src/interfaces/messageChat.ts
@@ -1,0 +1,6 @@
+export interface messageChat {
+    content: string;
+    createdAt: Date;
+    username: string;
+    id: string;
+}
